castone/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.0
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.7
numpy==1.26.4
gymnasium==0.29.1
pytest==8.2.2
httpx==0.27.0
PyJWT==2.8.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
websockets==16.0
pettingzoo==1.24.3
aiofiles==24.1.0
pytest-asyncio==0.23.7
torch==2.3.1
torchvision==0.18.1
